refactor(iota-indexer): consolidated new sub-modules hierarchy
- move pipelines into the ingestion module
- move top-level persist, transform & extract as sub-modules in the ingestion module
- removed unused structs and files as a result of the new module hierarchy
