+++ crates/iota-indexer/src/handlers/mod.rs
@@ -2,8 +2,6 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod checkpoint_handler;
-pub mod objects_snapshot_handler;
 pub mod optimistic_pruner;
 pub mod pruner;
 pub mod tx_processor;
