+++ crates/iota-indexer/src/ingestion/extract.rs
@@ -7,9 +7,15 @@
 use std::collections::BTreeMap;
 
 use iota_types::{
-    base_types::ObjectRef, digests::TransactionDigest, full_checkpoint_content::CheckpointData,
+    base_types::ObjectRef,
+    digests::TransactionDigest,
+    dynamic_field::{DynamicFieldInfo, DynamicFieldType},
+    full_checkpoint_content::CheckpointData,
     object::Object,
 };
+use move_core_types::language_storage::{StructTag, TypeTag};
+
+use crate::errors::IndexerResult;
 
 #[derive(Clone, Debug)]
 pub(crate) struct Extractor<'chk> {
@@ -49,4 +55,31 @@ impl<'chk> Extractor<'chk> {
         }
         eventually_removed_object_refs.into_values()
     }
+}
+
+/// If `o` is a dynamic `Field<K, V>`, determine whether it represents a Dynamic
+/// Field or a Dynamic Object Field based on its type.
+pub(crate) fn try_extract_df_kind(o: &Object) -> IndexerResult<Option<DynamicFieldType>> {
+    // Skip if not a move object
+    let Some(move_object) = o.data.try_as_move() else {
+        return Ok(None);
+    };
+
+    if !move_object.type_().is_dynamic_field() {
+        return Ok(None);
+    }
+
+    let type_: StructTag = move_object.type_().clone().into();
+    let [name, _] = type_.type_params.as_slice() else {
+        return Ok(None);
+    };
+
+    Ok(Some(
+        if matches!(name, TypeTag::Struct(s) if DynamicFieldInfo::is_dynamic_object_field_wrapper(s))
+        {
+            DynamicFieldType::DynamicObject
+        } else {
+            DynamicFieldType::DynamicField
+        },
+    ))
 }
