+++ crates/iota-indexer/src/ingestion/mod.rs
@@ -1,0 +1,8 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod extract;
+pub(crate) mod persist;
+pub(crate) mod primary;
+pub(crate) mod snapshot;
+pub mod transform;
