+++ crates/iota-indexer/src/ingestion/persist.rs
@@ -1,0 +1,144 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+//! Types and associated logic to use while persisting
+//! data to the database.
+
+use std::collections::BTreeMap;
+
+use async_trait::async_trait;
+use futures::{FutureExt, StreamExt};
+use tokio_util::sync::CancellationToken;
+use tracing::info;
+
+use crate::{errors::IndexerError, types::IndexerResult};
+
+pub(crate) const CHECKPOINT_COMMIT_BATCH_SIZE: usize = 100;
+pub(crate) const UNPROCESSED_CHECKPOINT_SIZE_LIMIT: usize = 1000;
+
+/// Defines the logic of writing operations to the database.
+///
+/// The writing can refer to one or multiple tables in the database.
+#[async_trait]
+pub trait Writer<T: Send + Sync + 'static>: Send + Sync {
+    /// Returns the writer name.
+    fn name(&self) -> String;
+
+    /// Commits batch of transformed data to DB.
+    async fn persist(&self, batch: Vec<T>) -> IndexerResult<()>;
+
+    /// Reads high watermark of the table DB.
+    async fn get_watermark_hi(&self) -> IndexerResult<Option<u64>>;
+
+    /// Sets high watermark of the table DB, also update metrics.
+    async fn set_watermark_hi(&self, watermark_hi: u64) -> IndexerResult<()>;
+
+    /// Gets the current max checkpoint that can be committed by the writer.
+    ///
+    /// This is for writers that have a predefined lag compared to the latest
+    /// checkpoint in the network.
+    ///
+    /// One use-case is the objects snapshot handler, which waits for the lag
+    /// between snapshot and latest checkpoint to reach a certain threshold.
+    ///
+    /// # Note
+    /// By default, returns `u64::MAX`, which means no extra waiting is needed
+    /// before committing.
+    async fn get_max_committable_checkpoint(&self) -> IndexerResult<u64> {
+        Ok(u64::MAX)
+    }
+
+    /// Processes the received data and persists it into a storage.
+    ///
+    /// - The data are received form the ingestion worker in which stage is
+    ///   transformed into something which can be directly commited into the
+    ///   database.
+    /// - The data received by this function are not guaranteed to be in order.
+    ///   The purpose of this function is to order the data by checkpoint
+    ///   sequence number and to ensure data committed are in order and
+    ///   contiguous.
+    ///
+    /// In addition, the method updates the watermark of the table of the data
+    /// is persisted to.
+    async fn persist_sequentially(
+        &self,
+        cp_receiver: iota_metrics::metered_channel::Receiver<(u64, T)>,
+        cancel: CancellationToken,
+    ) -> IndexerResult<()> {
+        let checkpoint_commit_batch_size = std::env::var("CHECKPOINT_COMMIT_BATCH_SIZE")
+            .ok()
+            .and_then(|val| val.parse().ok())
+            .unwrap_or(CHECKPOINT_COMMIT_BATCH_SIZE);
+        let mut stream = iota_metrics::metered_channel::ReceiverStream::new(cp_receiver)
+            .ready_chunks(checkpoint_commit_batch_size);
+
+        let mut unprocessed = BTreeMap::new();
+        let mut tuple_batch = vec![];
+        let mut next_cp_to_process = self
+            .get_watermark_hi()
+            .await?
+            .map(|n| n.saturating_add(1))
+            .unwrap_or_default();
+
+        loop {
+            if cancel.is_cancelled() {
+                info!("transform and load task terminating gracefully");
+                return Ok(());
+            }
+
+            // Try to fetch new data tuple from the stream
+            if unprocessed.len() >= UNPROCESSED_CHECKPOINT_SIZE_LIMIT {
+                tracing::debug!(
+                    "Unprocessed checkpoint size reached limit {UNPROCESSED_CHECKPOINT_SIZE_LIMIT}, skip reading from stream..."
+                );
+            } else {
+                // Try to fetch new data tuple from the stream
+                match stream.next().now_or_never() {
+                    Some(Some(tuple_chunk)) => {
+                        if cancel.is_cancelled() {
+                            info!("transform and load task terminating gracefully");
+                            return Ok(());
+                        }
+                        for (cp_seq, data) in tuple_chunk {
+                            unprocessed.insert(cp_seq, (cp_seq, data));
+                        }
+                    }
+                    Some(None) => break, // Stream has ended
+                    None => {}           // No new data tuple available right now
+                }
+            }
+
+            // Process unprocessed checkpoints, even no new checkpoints from stream
+            let checkpoint_lag_limiter = self.get_max_committable_checkpoint().await?;
+            while next_cp_to_process <= checkpoint_lag_limiter {
+                if let Some(data_tuple) = unprocessed.remove(&next_cp_to_process) {
+                    tuple_batch.push(data_tuple);
+                    next_cp_to_process += 1;
+                } else {
+                    break;
+                }
+            }
+
+            if !tuple_batch.is_empty() && checkpoint_lag_limiter != 0 {
+                let tuple_batch = std::mem::take(&mut tuple_batch);
+                let (last_checkpoint_seq, _data) = tuple_batch.last().unwrap();
+                let last_checkpoint_seq = last_checkpoint_seq.to_owned();
+                let batch = tuple_batch
+                    .into_iter()
+                    .map(|(_cp_seq, data)| data)
+                    .collect();
+                self.persist(batch).await.map_err(|e| {
+                    IndexerError::PostgresWrite(format!(
+                        "Failed to load transformed data into DB for handler {}: {e}",
+                        self.name()
+                    ))
+                })?;
+                self.set_watermark_hi(last_checkpoint_seq).await?;
+            }
+        }
+        Err(IndexerError::ChannelClosed(format!(
+            "Checkpoint channel is closed unexpectedly for handler {}",
+            self.name()
+        )))
+    }
+}
