+++ crates/iota-indexer/src/ingestion/snapshot/mod.rs
@@ -1,0 +1,5 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod extract_transform;
+pub(crate) mod persist;
