+++ crates/iota-indexer/src/store/pg_indexer_store.rs
@@ -26,6 +26,13 @@ use super::pg_partition_manager::{EpochPartitionData, PgPartitionManager};
 use crate::{
     db::ConnectionPool,
     errors::{Context, IndexerError, IndexerResult},
+    ingestion::{
+        primary::persist::{EpochToCommit, TransactionObjectChangesToCommit},
+        transform::{
+            CheckpointObjectChanges, LiveObject, RemovedObject,
+            retain_latest_objects_from_checkpoint_batch,
+        },
+    },
     insert_or_ignore_into,
     metrics::IndexerMetrics,
     models::{
@@ -44,9 +51,8 @@ use crate::{
         },
         tx_indices::TxIndexSplit,
     },
-    on_conflict_do_update, on_conflict_do_update_with_condition,
-    persist::{EpochToCommit, TransactionObjectChangesToCommit},
-    persist_chunk_into_table, persist_chunk_into_table_in_existing_connection, read_only_blocking,
+    on_conflict_do_update, on_conflict_do_update_with_condition, persist_chunk_into_table,
+    persist_chunk_into_table_in_existing_connection, read_only_blocking,
     schema::{
         chain_identifier, checkpoints, display, epochs, event_emit_module, event_emit_package,
         event_senders, event_struct_instantiation, event_struct_module, event_struct_name,
@@ -58,10 +64,6 @@ use crate::{
     },
     store::IndexerStore,
     transactional_blocking_with_retry,
-    transform::{
-        CheckpointObjectChanges, LiveObject, RemovedObject,
-        retain_latest_objects_from_checkpoint_batch,
-    },
     types::{
         EventIndex, IndexedCheckpoint, IndexedDeletedObject, IndexedEvent, IndexedObject,
         IndexedPackage, IndexedTransaction, TxIndex,
